refactor: move Leftpanel and Mainpanel outside repository
