+++ src/components/leftpanel/LeftPanel.sass
