+++ src/components/mainpanel/MainPanel.sass
